Move spec/login_helpers to spec/support/
