+++ spec/rails_helper.rb
@@ -6,7 +6,8 @@ abort("The Rails environment is running in production mode!") if Rails.env.produ
 require 'spec_helper'
 require 'rspec/rails'
 require 'capybara/rails'
-require 'login_helpers'
+require 'support/login_helpers'
+
 ActiveRecord::Migration.maintain_test_schema!
 
 RSpec.configure do |config|
